Remove CSS rule for hiding fieldset icons and lables
The original intend of that rule was not clear to me and it seems
it got dragged along for some time. I checked the DOM in the
collapsed state to see where this could even apply in the new
interface and found no other place than the one where we overwrite
it again.

I guess it can go away.

Change-Id: I2886db33c5f06d1c49acc4743c9acc198339de36