Fix 'File:' and 'Category:' links being piped unnecessarily
It is no longer necessary to prepend a colon in Parsoid HTML
to ensure they are interpreted as links rather than an image
inclusion or categorization.

Instead, the colon causes Parsoid to generate piped links
when they could be unpiped, so remove it.

This code was added in 1e62e9f64c (2012),
the Parsoid bug was fixed in b62b93c678 (2013).

Bug: T312700
Change-Id: I3d71fd658b5dd627445e60b850f647081ef842e7