Template dialog sidebar: Skip glitchy jQuery animation
What I found is a single `$container.animate( animations )` that's
responsible for the misbehavior described in T312768.

This method is called from a few places. I think none of these places
benefits from making this an animation. This is often more distracting
than helpful, especially when navigating with the keyboard.

Bug: T312768
Change-Id: I90a80d6ae8c1b47ee22297d2520255cad890b90e