Don't unnecessarily load huge icons for Wikimedia logos
The only icon we use from 'oojs-ui.styles.icons-wikimedia'
is the CC licenses logo.

Recently in I6217692e98088814c7a33bbc7ef8174622e03f25 more
Wikimedia project logos were added to that pack, and some
of them are positively huge. They may be improved in T328492,
but we're not using them anyway, so it makes sense to me
to use a custom icon pack here.

This reduces the download size for opening the editor for
the first time (uncached) by around 500 KB.

Change-Id: Ibe8a9ce9e4f1ded28fc9fd7e4ee41d5823123ceb