Fix trying to get a PageRecord for a non-existent page
This reverts a change from 5a792b6558. The hook only requires
a PageIdentity, and we only used toPageRecord() because it takes
a parameter to read latest data, and toPageIdentity() does not.
Instead call another method to update the internal data within the
Title instance. Maybe it's not great to rely on this side-effect…

Bug: T340568
Change-Id: If9cbd95eb84b6c39bf50af4564710839628a9bc3
(cherry picked from commit b986645d0ac3efadfabc9a32053c3e7e592fa294)