VisualEditorEdit diffs should only compare the main slot
We diff only on provided wikitext. This was causing inaccurate diffs on
pages with slotted content (e.g. File: pages on commons), which would
claim the edit would delete all the slot contents other than the
wikitext in main.

Also, remove the very old backwards compatibility params that were being
passed to getResultData in this method.

Bug: T351235
Change-Id: I135082162b96503e8d896aa4014abc9f4b936580