Fix type hint in ApiVisualEditorEdit
This doesn't make much of a difference, but I think it is more
correct. False is returned when the key cannot be found in the
cache. This is an expected situation here (hence the "try"). Null
can only be returned when null was stored in the cache before. This
can't happen here, as far as I understand the code.

Change-Id: I82f24e9f7234d0ec79f8223cd29e3df6e83f7f7b