EditCheck: Only pause if check is not rejected
The pause is to show the user the result of their
action (sepcifically it was designed for AddReferenceEditCheck
to let them see the reference in the page).

If they reject, there is no need to pause. This applies
to mulit-check too where it would cause a pause before moving on
to the next check.

Change-Id: I57f881b37051cc2e5ea6bda23fea66a2de2b342d