DesktopArticleTarget.init: don't clobber spans when updating single-tab
When in single tab mode with the pref set to remember-last, switching
your editor preference immediately updates the text of the edit links
on the page. The link HTML structure varies between different skins,
but we were assuming that the contents of the `<a>` could be replaced
with text. This lightly breaks rendering when the structure is a>span.
Instead, check first for a nested span, and fall back on just an a.

Follow-up to be36f1fab16d9adfbe8dce6946ecd2a10e55f30f

Bug: T389579
Change-Id: Ic323217e9d4325b52b34d67a3517cacbc51ac32a