Unbreak QuizGame compatibility
QuizGame (and some other extensions in the social tools family) intentionally creates an "edit" tab on a NS_SPECIAL page (Special:QuizGameHome).
The WikiPageFactory#newFromTitle call here fatals if $title points to a NS_SPECIAL page.
To work around that, let's just not call the method in question if we're on NS_SPECIAL.

Bug: T376487
Change-Id: I7c79c8ca1093ec01e7f516c66bdd5ed1d5057fd2