Migrate VisualEditor away from the Graphite PHP interface to Prometheus
This patch removes the use of PrefixingStatsdDataFactoryProxy under the
hood which makes use of the StatsdDataFactoryInterface and appends a key
prefix.

NOTE: No need to keep the old keys because Graphite is now read only.

This patch migrates to the new prometheus interface (stats lib) which is
now the recommended way today. In the future, I want to deprecate and
delete the PrefixingStatsdDataFactoryProxy class.

Change-Id: I7a61e965f5491f6686569f1192c7ebc5e0d3f262