DesktopArticleTarget: Fire wikitextInteractive hook after relevant reload
For some reason, some hooks (e.g. ve.activationComplete) are fired in
the Target class's code while others (e.g. ve.wikitextInteractive) are
fired in the first-activation script
(ve.init.mw.DesktopArticleTarget.init.js). I would've placed this change
in the former location if not for the latter. I suspect there to be a
design reason behind the latter's existence, so I chose to implement
this change without touching the latter.

Bug: T397778
Change-Id: If1a0faf60a7e12d22600b1d7372b76654ce8b072