Remove existing checks as an argument to edit checks
Checks received a copy of the existing checks so that they could look up
identical checks and consistently return the original version. This was
also used to allow state to be kept on a check. We no longer keep any
state on checks and we map to the existing checks in controller.js.

Bug: T397730
Change-Id: Ibcb9b49be18d5656b7b075c596557ae68b2b28b5