EditCheck: Add 'stale' state to checks
The stale check means that the check contents have been modified since the
check was last run.

A check can also be forced into a 'stale' state when the user clicks
an certain button e.g. 'revise' in ToneCheck.

The 'stale' state is shown the user with top and bottom borders of
a bounding rectangle.

Bug: T390250
Bug: T397984
Change-Id: I4ddd43395c616543c84e0c4fad5d8445207acbce