Edit check: set up the tone check a/b test
This changes the meaning of the VisualEditorEditCheckABTest config such
that it should be the name of a check in the experimental group, which
will be enabled on this wiki for users in the test-group.

The tone check is treated slightly specially because it's actually
loaded outside of the experimental module.

Bug: T389231
Bug: T402195
Change-Id: I9c195ff630a18be0a4e84799033f417185a3fcfe