Edit check: support checks looking for non-pure insertions
We wrote this to look for pure insertions because it was a requirement
for our initial add-reference check, which wanted to only trigger on
completely new content. This continued as the inherited behavior for
checks where it makes less sense. In particular, experimental checks
that are based on annotation changes don't function under this paradigm.

Changing this behavior allows the external link experimental check to
work when adding an external link to existing text.

Helpers/wrappers are added in order to preserve the existing behavior
for add reference.

Change-Id: I33e465ee2f20c4a9ae825038089df34f97b903bd