MobileArticleTarget: Simplify surface margin fix
This was to fix a scroll jump when loading the lede
section as described in T210630#4949865, but using a
simple margin apperas to work fine, for both the
lede and other sections, presumably as the read mode
CSS has changed enough since 2019.

This simplified CSS makes scrolling easier when the
"Edit full page" button is used.

Change-Id: I4172a899e54665e0cc17bc6b2aeb8cd10e1cb5d5