Paste URLs from Microsoft Edge address bar as plain text in source mode
Extend MWWikitextPlainTextStringTransferHandler to also handle the
'text/link-preview' format, similarly to how we handle URLs pasted as
plain text in source mode to not generate link markup.

Add integration tests for this in ClipboardHandler tests, as well as
integration tests for the aforementioned plain text URL behavior.
This requires a VE core patch to allow testing this case.

Depends-On: I3a39b577bbf33c49f5d2763d468b39a4761ffca8
Bug: T341281
Change-Id: I6565027dd5e5ec54510202efa66734cdeb165359