Close inspectors before switching to wikitext editor
In T296575 we discovered that both the reference as well as comment
inspectors (and possibly more) leave empty `<ref />` and `<!---->`
behind when switching editors. This doesn't happen when using the
save button.

The save button opens a dialog, which forces all inspectors to
close. Closing any inspector removes possibly dangling empty nodes
from the document. This patch enforces the same behavior when
switching editors.

This is a more universal solution compared to I6c8dc19 because this
works not only for references but all inspectors.

Bug: T296575
Change-Id: I1a3987f0c9bd98b1102477c7f80d131b491a7540